requirements.txt
matplotlib
nltk
numpy
Pillow
argparse
pycocotools
